Cleans up directories
