+++ src/api/aqsnv/auth.ts
@@ -1,15 +1,10 @@
 import config from "@/config/config";
 import {httpRequest} from "@/utils/http";
-import {AccessToken} from "@/types/auth";
 
 const endpoints = {
   pkce: () => new URL('/auth/pkce', config.api.aqsnv.server),
   authorize: () => new URL('/auth/authorize', config.api.aqsnv.server),
   token: () => new URL('auth/token', config.api.aqsnv.server),
-  me: () => new URL('/v1/profiles/me', config.api.aqsnv.server),
-  feedItems: () => new URL("/v1/feed/items", config.api.aqsnv.server),
-  feedItemAssignment: (feedItemId: string) => new URL(`/v1/feed/items/${feedItemId}/assignment`, config.api.aqsnv.server),
-  feedItemCompletion: (feedItemId: string) => new URL(`/v1/feed/items/${feedItemId}/completion`, config.api.aqsnv.server),
 };
 
 export type PKCEPair = {
@@ -52,84 +47,4 @@ export async function exchangeAuthorizationCode(authorizationCode: string, verif
     method: "POST",
   });
   return response.json();
-}
-
-export type User = {
-  name: string,
-  pictureUrl: string,
-};
-
-export async function getCurrentUser(token: AccessToken): Promise<User> {
-  const response = await httpRequest(endpoints.me(), {
-    headers: {
-      "Authorization": token.asAuthorizationHeader(),
-    },
-  });
-
-  return response.json();
-}
-
-export type FeedItem = object;
-
-export type Feed = {
-  backlog: [FeedItem],
-  inProgress: [FeedItem],
-  done: [FeedItem],
-};
-
-export async function getFeedItems(token: AccessToken, status: "backlog" | "inProgress" | "done") {
-  const url = new URL(endpoints.feedItems());
-  url.searchParams.append("status", status);
-  url.searchParams.append("limit", "5");
-  const response = await httpRequest(url, {
-    headers: {
-      "Authorization": token.asAuthorizationHeader(),
-    },
-  });
-
-  return response.json();
-}
-
-export async function assignFeedItem(token: AccessToken, feedItemId: string): Promise<void> {
-  await httpRequest(endpoints.feedItemAssignment(feedItemId), {
-    method: 'post',
-    headers: {
-      "Authorization": token.asAuthorizationHeader(),
-    },
-  });
-
-  return;
-}
-
-export async function unassignFeedItem(token: AccessToken, feedItemId: string): Promise<void> {
-  await httpRequest(endpoints.feedItemAssignment(feedItemId), {
-    method: 'delete',
-    headers: {
-      "Authorization": token.asAuthorizationHeader(),
-    },
-  });
-
-  return;
-}
-
-export async function completeFeedItem(token: AccessToken, feedItemId: string): Promise<void> {
-  await httpRequest(endpoints.feedItemCompletion(feedItemId), {
-    method: 'post',
-    headers: {
-      "Authorization": token.asAuthorizationHeader(),
-    },
-  });
-
-  return;
-}
-
-export async function uncompleteFeedItem(token: AccessToken, feedItemId: string): Promise<void> {
-  await httpRequest(endpoints.feedItemCompletion(feedItemId), {
-    method: 'delete',
-    headers: {
-      "Authorization": token.asAuthorizationHeader(),
-    },
-  });
-
-  return;
 }
