--- src/routes/Landing.tsx
@@ -1,10 +1,0 @@
-import {RequiresAuthorization} from "@/hooks/auth";
-import Layout from "@/Layout";
-
-export function Landing() {
-  return (
-    <RequiresAuthorization>
-      <Layout />
-    </RequiresAuthorization>
-  );
-}
