+++ src/routes/index.tsx
@@ -1,5 +1,5 @@
 import {BrowserRouter, Routes, Route} from 'react-router-dom';
-import {Landing} from './Landing';
+import {Layout} from '@/components/Layout';
 import {Feed} from './Feed';
 import {AuthorizationCallback} from '@/hooks/auth';
 
@@ -8,7 +8,7 @@ export function Router() {
     <BrowserRouter>
       <Routes>
         <Route path='/oauth/cb' element={<AuthorizationCallback />} />
-        <Route path='/' element={<Landing />}>
+        <Route path='/' element={<Layout />}>
           <Route path='/feed' element={<Feed />} />
         </Route>
       </Routes>
